app-text/ghostscript-gpl: allow building against libpng-1.2
Chrome uses the libpng-1.2.x series, and we've had problems upgrading
our libpng version for Chrome OS, so we've pinned libpng to <1.3. This
causes conflicts with the current dependencies for ghostscript.

The ghostscript sources and docs never mention requiring libpng >=
1.6.2, and I have several Ubuntu systems that have ghostscript linked
just fine against libpng-1.2.x, so I don't believe it's a technical
limitation. And now, I can build and run ghostscript just fine, AFAICT.

It appears, then, that the dependency was chosen for the upstream ebuild
was arbitrary, so let's change this to >=1.2.

BUG=chrome-os-partner:47758,chrome-os-partner:49267
TEST=`emerge-${BOARD} ghostscript-gpl` works without unmasking >=libpng-1.3
TEST=run CUPS + a few PostScript printers

Change-Id: I6e25d67178bc57b984e9ed0decf8de94e45a4ca5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/322293
Reviewed-by: Mike Frysinger <vapier@chromium.org>