dev-lang/go: upgrade to go-1.6.
This ebuild brings Go 1.6 to Chrome OS (both host and cross compilers).
For details on testing and known issues, see http://go/Go1.6-On-ChromeOS

Most of the changes are in the implementation of the language, runtime,
and libraries. There are no changes to the language specification.
For more details, see https://golang.org/doc/go1.6

BUG=None
TEST='emerge cross-x86_64-cros-linux-gnu/go' and ran compiler and standard library tests on a panther device.
TEST='emerge cross-i686-pc-linux-gnu/go' and ran compiler and standard library tests on a x86-zgb device.
TEST='emerge cross-armv7a-cros-linux-gnueabi/go' and ran compiler and standard library tests on a daisy device.

Change-Id: Ifaf09219895fefb831ff8dcc2e2d3310201c62c2
Reviewed-on: https://chromium-review.googlesource.com/324842
Commit-Ready: Rahul Chaudhry <rahulchaudhry@chromium.org>
Tested-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>