net-print/cups: add init script
CUPS needs some specific directories and permissions to execute
properly. It also wants to be able to read/write printer configuration
(printers.conf and ppd/*.ppd) into its PrinterRoot directory
(/var/cache/cups/printers) [1]. So we check (and create/chown/chmod, as
needed) most of the stateful directories on the fly.

Don't handle any session management (e.g., autostart, or halt on
logout) yet. We'll look into starting CUPS based on udev rules (for USB)
or a printer management UI (for network printing), so no 'start on'
rules for now. And we may need to hook into the session manager for
handling log out and/or multiprofile.

When minijailing cupsd, let's include a few extra flags, to run in a new
PID, IPC, and mount namespace (but not network namespace, as we want to
be able to reach network printers), to set no_new_privs, and to remount
/proc read-only. AFAICT, CUPS should operate fine under those
constraints.

BUG=chrome-os-partner:47758,chrome-os-partner:49267
TEST=`start cupsd`
TEST=After starting cupsd, run something like the following as chronos:
     wget -O /tmp/Ricoh.ppd 'http://www.openprinting.org/ppd-o-matic.php?driver=Postscript-Ricoh&printer=Ricoh-Aficio_SP_C320DN&show=0'
     wget -O /tmp/test-pdf.pdf 'http://www.orimi.com/pdf-test.pdf'
     lpadmin -p RICOH_IPP_3 -P /tmp/Ricoh.ppd -E -v "ipp://192.168.1.3/printer"
     lpr -P RICOH_IPP_3 /tmp/test-pdf.pdf

Change-Id: I3a38c039ee8471cae20128819d9bc125a7bd4d17
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/328012
Reviewed-by: Mike Frysinger <vapier@chromium.org>