move nsswitch.conf from shill into its own ebuild
Do this because:
(a) I need to provide an alternate conf file when enabling mdns lookup
    support and
(b) this knowledge doesn't need to be coordinated in shill; it should
    stay in CrOS ebuilds

Regarding the nsswitch.mdns.conf selections:

Only use the '_minimal' version, due to suggestions by the upstream
developer [0], due to issues seen with (e.g.) SSH on various other
distros. This provides lookups for local addresses only (which is most
of what mDNS is used for).

Also, do only IPv4 lookups. See excerpt from the README [1]:

  "Due to the fact that most mDNS responders only register local IPv4
   addresses via mDNS, most people will want to use libnss_mdns4.so.2
   exclusively. Using libnss_mdns.so.2 or libnss_mdns6.so.2 in such a
   situation causes long timeouts when resolving hosts since most modern
   Unix/Linux applications check for IPv6 addresses first, followed by a
   lookup for IPv4."

[0] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=433943
[1] http://0pointer.de/lennart/projects/nss-mdns/

BUG=chromium:588234
TEST=`USE=zeroconf emerge-${BOARD} shill nss-mdns`; `start avahi; getent hosts HOSTNAME.local`
TEST=make sure regular DNS lookups still work

Change-Id: I5b1a515e50b8bded41a6ffd2e15e1309c647f698
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/328768
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>