cros-firmware: Add PD image to local firmware updater.sh
This change will add a PD firmware image to a local build of the
firmware updater.sh if it exists in the firmware directory.  This
is useful for doing testing of the updater process by using local
updater images instead of needing to build official images.

BUG=chromium:577269
TEST=emerge chromeos-firmware-chell and check that updater.sh
also includes a PD firwmare image.

Change-Id: I98082fdb2ed8266c5d3c1d2f39a3b0949bb3e8a2
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329230
Reviewed-by: Aaron Durbin <adurbin@chromium.org>