Force applying of fix for chromium:593719
This puts the fix for this bug in the current ebuild.
The fix is in https://chromium-review.googlesource.com/332221.
We need this fix urgently and we cannot wait for this change to be
uprevved by the PFQ. So, I am applying directly to the current ebuild.

BUG=chromium:593719
TEST=tested previously. Just started an emerge by hand to make sure I
     did not break anything.

Change-Id: I3f23621daa7a6e0f95127771ef498a615a847015
Reviewed-on: https://chromium-review.googlesource.com/332296
Reviewed-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>