chromeos-accelerometer-init: Run only the necessary commands
The current script tries sysfs entries for old and new accelerometer layout.
By testing the presence of file "location", we only issue the right commands.
Also, load sysfs trigger module if necessary. It is required for triggering
from sysfs, but not by the sensors.

BUG=chrome-os-partner:50851
TEST=Check it works on Cyan and Samus.
Check with sh -x and udev.conf debug logging that now error messages are
issued.

Change-Id: Ic8c4687a2da265da1d90f0ac5f034dc2b57b7102
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332437