chromeos-accelerometer-init: Make udev grant kb wake access to powerd
This patch extends the udev init script for the EC accelerometer driver
to grant write access of the in_angl_offset or kb_wake_angle node (one
of which controls the motionsense kb_wake property on the EC, depending
on kernel version) to the power group. It is used by powerd to control
wakeup behavior.

BUG=chromium:594037
TEST=Deployed on Minnie and confirmed the access rights were changed
accordingly after reboot.

Change-Id: I59f0ed2b4c9c57388355729d8608cdde870644e4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/333075
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>