userfeedback: Add dependency on crash-reporter
CL:334910 adds functionality to crash-reporter that uses the
/var/log/eventlog.txt file generated by userfeedback. Since it's not
desired to add a dependency from crash-reporter on userfeedback,
I've instead added a dependency from userfeedback on crash-reporter
(by changing the userfeedback init script in a way that will make
crash-reporter-late wait on it iff it exists). Update the userfeedback
ebuild to reflect this new dependency.

CQ-DEPEND=CL:334910
BUG=chromium:595531
TEST=None

Change-Id: I190985e3b82d378119bcc2fc4707d3e740d7394e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/335264