net-wireless/bluez: set correct mouse properties
Microsoft Sculpt Touch Mouse advertises itself as a bluetooth
mouse. But it reports the capability of sending keyboard events
as well as mouse events. And hence udev adds this mouse into the
input subsystem with ID_INPUT_KEYBOARD=1 and ID_INPUT_MOUSE=1.
The incorrect properties make chrome OOBE consider this mouse
as a keyboard/mouse combo and prevent other real bluetooth
keyboard from pairing.

This patch uses dbus to query the bluetooth class of device
(CoD). If the CoD indicates that the device is a mouse, the
keyboard related properties in the input subsystem are reset.

BUG=chromium:570005
TEST=Use a Microsoft Sculpt Touch Mouse on Mickey.
Boot into OOBE of Mickey.
Observe that it is paired as a mouse only.
It is not paired as a keyboard.
And confirm that another real bluetooth keyboard could pair
without problem.

Change-Id: I0cb810a5306c29ae635ddde06c6121cb49447240
Reviewed-on: https://chromium-review.googlesource.com/336598
Commit-Ready: Miao-chen Chou <mcchou@chromium.org>
Tested-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>