[Reland] sshd: try to make ssh connection more reliable.
Provisioning a new OS image shows many instances of ssh
connections droped, like
  autoserv|  Master ssh connection to w.x.y.z is down.
  autoserv| Starting master ssh connection

Many of these are harmless. But even manually connecting to a DUT
I get dropped connections, I think when the machine gets busy,
but hasn't rebooted yet:
  packet_write_wait: Connection to w.x.y.z: Broken pipe

Autotest currently connects to a DUT with the following options:
  ConnectTimeout=30
  ServerAliveInterval=900
  ServerAliveCountMax=3
  ConnectionAttempts=4
  Protocol=2

Basically it puts the burden of keeping the connection alive on the
server controlling the DUT. Interestingly we do see failures after
900 seconds quite often, sometimes after 1350s. As the sampling
is so coarse it is hard to say if this is just bad luck, bad
network or a rebooted/dead machine.

This change tries to make the DUT more cooperative by letting it
ping the connection to the ssh client (autotest server) once a
minute, ignoring all network errors for half a day (one hour would
probably be enough but I see no harm in being generous). It does not
change how autotest itself connects to the DUT.

BUG=chromium:598517
TEST=Made a lars image, ssh to it.

Change-Id: I1571a0b19d1ec2eb164151b36c11d0f73536d548
Reviewed-on: https://chromium-review.googlesource.com/338898
Commit-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>