chromeos-chrome: switch 32 bit to component build.
We may need to revert this change once we have a better fix.

We don't have good symbols on the crash server. This breaks the
chrome binary into multiple libraries. Unfortunately it uses
more space on disk and probably will be slower in benchmarks,
but the latter should be checked across boards in the lab.

BUG=chromium:595763, chromium:601854
TEST=emerge-lars chromeos-chrome
     emerge-daisy_skate chromeos-chrome

Size on daisy goes from
  100M chrome
to
   41M chrome
   96M lib

Run benchmarks on daisy:
Sunspider
from
  917.9ms +/- 1.8%
  929.6ms +/- 1.1%
to
  924.3ms +/- 1.5%
  928.8ms +/- 2.1%

Kraken
from
  4880.6ms +/- 4.4%
  4854.5ms +/- 3.2%
to
  4889.3ms +/- 3.3%
  4854.5ms +/- 3.6%

WebGL Aquarium different fishes (benchmark might be broken right now)
50    from random 30..50 to stable 50
1000  from 28..31 to 25..27
4000  from  9..11 to 9..11

Change-Id: Iba55a3f760100c330ec16e2db906b39f521517d7
Reviewed-on: https://chromium-review.googlesource.com/340057
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>