media-libs/mesa: Removed code that causes Mesa to return a desktop GL enum
When the draw_buffer_enum_to_bitmask function
is called while Mesa is in it's meta state,
the wrong GLbitfield is returned sometimes for
OpenGL ES contexts.  This is because the
_mesa_meta_begin function makes the ctx->API always
equal to desktop GL during the meta state.

In the case of GL_BACK, the _mesa_is_gles(ctx)
function is checked when returning the enum.
This check evaluates to false during the meta state,
even on mobile devices.

BUG=b/27656575, chromium:606472
TEST=Ran dEQP-GLES2.functional.clipping.line.line_attrib_clip
with the pbuffer option with on Cyan.  Verified it
fails without the patch.

Change-Id: I111fc61dfeb1f8289ab71227c9ac9b7e6612c002
Reviewed-on: https://chromium-review.googlesource.com/340347
Commit-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>