media-libs/mesa-img: Remove code that causes Mesa to return a desktop GL enum
When the draw_buffer_enum_to_bitmask function
is called while Mesa is in it's meta state,
the wrong GLbitfield is returned sometimes for
OpenGL ES contexts.  This is because the
_mesa_meta_begin function makes the ctx->API always
equal to desktop GL during the meta state.

In the case of GL_BACK, the _mesa_is_gles(ctx)
function is checked when returning the enum.
This check evaluates to false during the meta state,
even on mobile devices.

This syncs mesa-img with mesa @ 79e3f70

BUG=chrome-os-partner:52663
TEST=WebGL aquarium.  dEQP tests as discussed in the bug.

Change-Id: I8d26999bb0145e48add5b33fb4ae1186f2bc68d2
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/342962
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Mao Huang <littlecvr@chromium.org>