Faster sync Chrome in distdir, build in the cache dir
I was profiling a build time issue on Cyan, and found out that the
operations past building Chrome added about twenty minutes to the build
time. This was causing timeout issues on some of the builders.

I found that the run time in particular of
install_telemetry_dep_resources() was almost all of the time, consisting
of about half a minute to run a python script to get a list of about
twelve thousand dependencies, followed by more than fifteen minutes to
copy each file independently.

Examining what install_test_resources I observed it used "rsync" to copy
each individual file (though some inputs could be directories from other
callers) into a cache directory, and then a subsequent "cp -al" to
create a hard link from there into a destination directory.

The replacement code leverages rsync arguments to take a list of
relative paths to copy, which is effectively what the input file list
was. The list is now written to a temporary file (set up to be
automatically deleted), and rsync is invoked twice. Firstly to copy all
of the files to the cache directory, and then secondly to generate the
hard links.

There is also a minor fix at the point of call to quote DEPS_LIST, as
some of the files being copied have spaces in their names. The previous
code would generate warnings about files not being found for the copy,
and now those files are correctly handled.

The replacement code takes about two seconds, so the total time
of install_telemetry_dep_resources() shrinks the thirty seconds to get
the list of python dependencies. I'm sure that minute can be shrunk too,
but I did not attempt it.

With the other post-build operations otherwise profiling as taking
just seconds, this should be a good speed up for the builders.

TEST=rm'd the cache and dest directories
TEST=verified that they are recreated and hard-linked
TEST="emerge-${BOARD} chromeos-base/chromeos-chrome" is faster
BUG=chromium:610807

Change-Id: Ie83568bb415430c9d85669b842186f4484114006
Reviewed-on: https://chromium-review.googlesource.com/344931
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Trybot-Ready: Lloyd Pique <lpique@google.com>