libpayload: Minimalize config files
It used to be that libpayload board config files needed to contain
exactly every set and unset option relevant for a board to make it build
correctly in all situations, but that restriction is long gone. 90% of
the content of our configuration files is now useless boilerplate about
options that aren't enabled or options that are part of the defaults
anyway.

This patch replaces all configs with the actually relevant minimum, by
regenerating them with:

	cd <...>/coreboot/payloads/libpayload
	for c in <...>/sys-boot/libpayload/files/configs/*; do
		cp $c .config
		make savedefconfig
		mv defconfig $c
	done

Also remove some redundant nyan_xxx and veyron_xxx configs since the
ebuild will nowadays just pick up the file for the base board if there
is none for the derivative.

CQ-DEPEND=CL:347970
BUG=None
TEST=Built and booted Jerry and Oak. Reexpanded all configs with 'yes ""
| make oldconfig' and compared to old versions, manually confirmed that
there were no differences except for those from previous
inconsistencies that hadn't been renormalized.

Change-Id: Ib8a9da52b6b889a5c12e80767f555893bfdf3753
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/344950
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>