Fix install_test_resources for directories
The switch to using rsync with a list of paths rather than having a loop
that copied each one individually with rsync introduced an unfortunate
regression. The previous code copied the contents of a directory, the
replacement did not. I thought I had tested that possibility, but a bug
was reported because it was not.

This patch fixes it -- when using --files-from with -a, you also need to
explicitly include -r to get recursive behavior.  This behavior is
documented in the man page.

BUG=chromium:614925
TEST=Verified contents of media/test/data are copied
TEST=Verified also that individual files are copied
TEST=install_telemetry_dep_resources was not significantly slower

Change-Id: Icc2e49cc994719a463b46768ddeef41a1f6222bf
Reviewed-on: https://chromium-review.googlesource.com/347474
Tested-by: Lloyd Pique <lpique@google.com>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Lloyd Pique <lpique@google.com>