Pass -Wno-unknown-warning-option option only to Clang.
I added -Wno-unknown-warning-option a week or so ago.
I thought this was a valid option for both Clang and GCC. It turns out it
is only supported by Clang. GCC does not support it but it will only
complain about it if there are other errors in the file.

GCC will only complain about it if you pass it in the "positive" form.
ie: -Wunknown-warning-option
$x86_64-pc-linux-gnu-gcc -Wunknown-warning-option /tmp/test.c
x86_64-pc-linux-gnu-gcc.real: error: unrecognized command line option -Wunknown-warning-option

And it will not complain in the negative form if there are no
other errors in the file:
$ x86_64-pc-linux-gnu-gcc -Wno-unknown-warning-option /tmp/test.c
$ echo $?
0

BUG=chromium:615466
TEST=cbuildbot falco-release

Change-Id: I0b09415236b7a3edd985229c221fcdcb7325602f
Reviewed-on: https://chromium-review.googlesource.com/347699
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>