[GCC] Don't mix -Wl,--build-id with -r flags.
The compiler wrapper recently added -W,--build-id.  This works most
of the time, but it should not be added if the -r or --relocatable
flag is passed to the compiler.  We have been seeing a few files with
multiple build ids as a result of the recent change.  This fixes that.

BUG=chromium:616559
TEST=Tested in my chroot.

Change-Id: Ia2a59912e8e3bc17652861120a622fc76d99f222
Reviewed-on: https://chromium-review.googlesource.com/348880
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>