[llvm] Configure for llvm to pass --build-id to linker.
GCC is automatically configured to pass --build-id to the linker
when the linker is called, unless the linker.

Upstream LLVM was recently updated so that it can be configured to work
just like GCC does for this. This CL backports those changes to our
version of llvm.  It also removes a failed attempt to add this
functionality in the compiler wrapper (that was causing the linker to
always be invoked, which caused some errors).

BUG=b:28897967
TEST=Tested in my chroot. Successfully tested on trybots:
https://uberchromegw.corp.google.com/i/chromiumos.tryserver/builders/release/builds/4226
https://uberchromegw.corp.google.com/i/chromiumos.tryserver/builders/release/builds/4225
https://uberchromegw.corp.google.com/i/chromiumos.tryserver/builders/release/builds/4227

Change-Id: I04ad285332bcad04d05ce9ff287cb8a5fb44a104
Reviewed-on: https://chromium-review.googlesource.com/349721
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>