media-libs/mesa-img: Add a DRI Query Buffers extension
The extension is used to notify the DRI module that it should fetch
updated buffer information immediately. This fixes some failures in
the dEQP EGL surface resize tests, where the surface is resized
immediately before calling eglSwapBuffers, and the surface
dimensions queried immediately afterwards (using eglQuerySurface).
Without the extension, the buffer information is updated when
KEGLGetDrawableParameters is next called, but this is too late for
the test cases in question.

This should provide the necessary bits to fix the bug below on
ARC, but we will need to make the appropriate changes to
platform_android.c

It has been tested on X11 and Wayland

CQ-DEPEND=CL:*262435
BUG=chrome-os-partner:54012
TEST=Boot to UI.

Change-Id: Ia13114d5791ca2b2888298cd060222baaa92db92
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/349600
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>