cros-kernel2.eclass: detect and unset incompatible CC
tc-export doesn't change environment variables when they are set
already. When CC is customly defined, it is used to build both
userland programs and the kernel. That is a problem if they need
different compilers.

For example, when CC is set to armv7a-cros-linux-gnueabi-gcc for
userland programs, it will also be used to compile the 64-bit
kernel, which requires aarch64-cros-linux-gnu-gcc.

This patch unsets incompatible CC when necessary so that tc-export
can set CC for kernel properly.

TEST=built aarch64 kernel with profile llvm
BUG=chromium:619947

Change-Id: I8e5c78fc41aa0d758416a0a60a5a06b848d460d4
Reviewed-on: https://chromium-review.googlesource.com/352490
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Ting-Yuan Huang <laszio@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>