net-print/cups: start using seccomp for cupsd
I ran cupsd through `strace -f`, with a few test invocations for
printing, and came up with this list of necessary syscalls. I ran on an
ARM32 system only for now, and whittled the trace down with some
sed/grep and `sort -u`. I had to do the following translations for ARM,
since apparently strace is a bit *too* user-friendly at converting
syscall names for us (whereas minijail is not):

  fadvise64_64 -> arm_fadvise64_64
  getrlimit -> ugetrlimit
  pread -> pread64
  select -> _newselect
  set_tls -> ARM_set_tls

amd64 needed the following translation:

  _exit -> exit
  pread -> pread64

x86 was run in qemu, and didn't get as thorough of a test; I also merged
in some syscalls that were present in the amd64 run, but not x86.

BUG=chromium:620192
TEST=`test_that platform_CUPSDaemon`, and other simple CUPS tests

Change-Id: Id83d148a3fc86a2fb04ab0bfa94003c6c31014b3
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/352761
Tested-by: Sean Kau <skau@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Justin Carlson <justincarlson@chromium.org>