app-text/ghostscript-gpl: resolve parallel 'make install' issues
Gentoo developers have previously noticed some dependency bugs that
caused parallel builds to fail:

  https://bugs.gentoo.org/show_bug.cgi?id=550926

I haven't seen these problems, but it appears to be the
directory-creation dependency that was fixed here by adding an ordering
dependency on the 'make directories' target:

  http://bugs.ghostscript.com/show_bug.cgi?id=696271
  http://git.ghostscript.com/?p=ghostpdl.git;a=commitdiff;h=e126995d6327788ddac7fd99f55db3c1603beea7
  e126995d6327 "Bug 696271: Fix a load of makefile dependencies."

That fix was included in 9.19.

I also reported and fixed a parallel 'make install' bug here:

  http://bugs.ghostscript.com/show_bug.cgi?id=696843

Let's apply this fix, since I've observed it.

Anyway, the build takes forever on a single thread (upwards of 5
minutes). I've stress tested this in a loop now, so let's undo the -j1
restriction.

BUG=chromium:588232
TEST=`while : ; do emerge-${BOARD} ghostscript-gpl || break; done`

Change-Id: I8eff1de803289a05738d7236dedc6d5d2abb474c
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353035
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>