net-print/cups: add 'lp' user/group back to ebuild
The lp user/group aren't very useful to us, but the group is used as the
means for cupsd to get permissions over USB printers. I previously
dropped 'lp' from our custom ebuild, but I don't see a good reason to
continue to do this. This user/group also happens to get setup in the
chromeos-base ebuild, so we didn't notice that we still needed this.

BUG=chromium:616866
TEST=build

Change-Id: I6cc3c018deadeb6fc7d39b5484ea44dbb8da2eb6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/353154
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>