mesa: correct the git_sha1.h step for the 9999 ebuild
It turns out that the cwd of src_prepare() is not a git checkout, so the
'git log' command won't work, and it isn't actually possible to
determine the git hash. Instead we can just produce a dummy git_sha1.h
header file to make the build work.

The 9999 ebuild is for dev testing only, and as such we really do not
the accurate git hash.

BUG=chromium:620830
TEST=emerge-cyan-cheets mesa and check the portage work directory

Change-Id: I3456a18f2cb0fdddb292dcbb4d0159ac2da0c0f8
Reviewed-on: https://chromium-review.googlesource.com/353720
Commit-Ready: Haixia Shi <hshi@chromium.org>
Tested-by: Haixia Shi <hshi@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>