mariadb: work around cros copmiliation error.
New cmake_utils.eclass detects the build of mariadb on the guado-moblab
board is cros build. It disables the build of some binary that runs on the host,
this causes the build failure. This CL workarounds it by letting cros
compiler build binary that runs on the host. This is the old/working
behavior.

BUG=chromium:621558
TEST=ebuild-guado-moblab mariadb passes on upstream cmake-utils.eclass

Change-Id: I6cc7ddf36bb7fa9762627a8b1a2ed34db27206b4
Reviewed-on: https://chromium-review.googlesource.com/354692
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>