disable mDNS resolution for glibc
This conflicts with resolution of the non-standard .local TLD used in
some enterprises. It isn't yet critical to any other major launched
features, so let's disable it for now while we figure out a more robust
plan for when/if we want to turn this back on.

This is a partial revert of commit 194e3fbc9b6d ("turn on mDNS
resolution in ChromeOS").

BUG=chromium:611631
TEST=check that mDNS names can no longer be resolved (e.g., find nearby
     device via `avahi-browse -atr` and try to `ping foo.local`)
TEST=resolve foo.local hosts from a local DNS server

Change-Id: I25b67f9554aeb6ffd95746068dd9f776c32fb3d1
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/355140
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>