sysroot_wrapper: use relative location for clang compiler
When the wrapper is used for clang in simple_chrome, the wrapper expects
the user to set the location of clang with envvar CLANG.
However, the goma wrapper does not propagate the value of this envvar to
the sysroot_wrapper.
To avoid this problem, we change the default location to be relative
to the location of the sysroot wrapper which is valid both within the
chroot and in the simple chrome environment.

BUG=chromium:625060
TEST=cbuildbot --latest-toolchain daisy-release falco-release

Change-Id: I4a3f7cfc4a4fb4bc2b92830b892b29c979e25ae2
Reviewed-on: https://chromium-review.googlesource.com/357823
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>