media-libs/waffle: Set correct surface type
ChromeOS doesn't have native windows, Waffle assumes it does. Hardcode
the surface type here so we can drop a hack currently in Mesa (Waffle
hacks are easier to maintain!).  A more correct way to fix this would
be to pass in the surface type to the wflinfo command line, but that
would require changes in every test that uses wflinfo.

BUG=chromium:606472
TEST=Ran:

wflinfo -p null -a gles2

Verified it worked on cyan (with Mesa hack removed) and veyron_minnie.

Change-Id: I0d74358d88802699d588b7db94b0d0f3ffc7d053
Reviewed-on: https://chromium-review.googlesource.com/358936
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>