chromeos-fonts: Pre-cache fonts
Building the chromeos-fonts package is slow, taking as long as 9 minutes
to build on ARM boards. This is because we run fc-cache under qemu at
install time and it takes an obscenely long time for fc-cache to process
the notocjk fonts.

For generic linux distributions, the font cache needs to be generated at
install time because there is no way to know what fonts are installed on
the system or in the user's directory.  We don't have this problem on
chrome os since we control exactly which fonts are installed in the
rootfs and users don't have the ability to install additional fonts.

We can take advantage of this to generate the font cache once on a
builder and then distribute the cache files as part of a pre-built
package so that each developer doesn't need to manually generate the
exact same font cache on their machine.

This CL makes the changes necessary to make this happen:

- /etc/fonts/fonts.conf now doesn't list any font directories.  Instead
  all the fonts are explicitly listed in the local.conf file in that
  directory instead.  The fonts are explicitly listed to prevent random
  other directories that may be floating around in the sysroot from
  causing fontconfig to think that the cache is invalid when it checks
  it at boot time on the running system.  However, this does mean that
  a new font needs to be explicitly added to the local.conf file before
  it will be included in the rootfs (in addition to any other steps
  necessary to get the font installed).  On the other hand, new fonts
  are rarely added and the current situation is making everyone who
  develops for ARM boards miserable.
- fontconfig conveniently stores the mtime of the font directory in the
  cache file for that font.  Since mtimes are not preserved when
  installing from a binary package this means that the mtime for the
  font directory will always be different from the one in the cache.
  To deal with this, we remove the directory mtime check from
  fontconfig.  Fonts cannot be added, removed, or changed at runtime so
  the mtime check doesn't actually buy us anything.
- fc-cache runs during the compile phase of the ebuild.  The two issues
  with this are that it uses sudo to chroot into the board's ROOT and
  that it references ROOT during the compile phase at all.  Both are
  considered bad by gentoo but they might be ok in this case since
  chromeos-fonts is such a specialized package and we get such a
  significant speedup from doing it.
- We skip generating the font cache for the chroot.  We don't need this
  cache in the chroot and this matches the previous behavior of the
  ebuild where it only built the font cache if cros_target ==
  target_image.

Overall this change speeds up build_image for ARM devices by 4 to 9
minutes.

BUG=chromium:205424
TEST='emerge-kevin chromeos-fonts' completes successfully

Change-Id: I0c16fe48bfc3abaf0702daf754ca1f5163d7c82d
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/428177