coreboot/libpayload: Clean up configs for Gru and Kevin
This patch removes some redundant or incorrect Kconfig options from Gru
and Kevin config files which are already part of the defaults anyway.
Also add the OHCI host controller driver to Gru because its depthcharge
code is actually trying to use that. Alphabetized Kevin's coreboot
config to put it in line with Gru.

Added the deselection line for CONFIG_CONSOLE_SERIAL to both boards to
make our build system (with the fwserial addition for developer but not
production builds) work, because otherwise the serial console is by
default enabled in all images.

CQ-DEPEND=CL:361581
BUG=None
TEST=Built Kevin image.bin, confirmed that it booted and had no serial
output.

Change-Id: I6293887f7123c6885d0f396b33da9166d7c230c5
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/361571
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>