media-libs/mesa-img: Garbage collect sync objects more often
When onscreen keyboard is visible, Chrome creates and destroys a lot of
EGLSync objects.  If the sync is destroyed before it is triggered, then
we have to defer the cleanup.  Prior to this patch, we were deferring
the cleanup until app shutdown.  But that meant we were effectively
never cleaning up the syncs.

Fix is to do the cleanup more regularly.

BUG=chrome-os-partner:55035
TEST=Modify img-ddk to show number of outstanding syncs
     Run browser with on-screen keyboard.
     Make sure number of outstanding syncs stays reasonable
     Also general web browsing and youtube videos.

Change-Id: I430c8d308ba285218400c66a16dd7a5ab64cc1a1
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/362551
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Horng-Shyang Liao <hs.liao@mediatek.com>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>