[gdb] Patch GDB with patch from upstream to fix remote packet issue.
This applies the patch from
https://sourceware.org/bugzilla/show_bug.cgi?id=13984
to fix an issue with remote kernel debugging with aarch64.  The
patch has not been accepted upstream, which is why we apply it
here rather than in the actual gdb source repository.

BUG=chromium:630015
TEST=Built gdb with this CL; gave patched GDB binary to person
with issue and it fixed the problem.

Change-Id: Ib014858c9d3fbe6de0a67bc13db5445e8251dc0c
Reviewed-on: https://chromium-review.googlesource.com/362984
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>