ARC: Add firewall rule to accept incoming ssdp traffic
Some UPnP-enabled apps use SSDP for discovery of devices
on the local network. In order for such broadcast messages
to reach the container, we need to explictly accept such
traffic so that it's not dropped by the firewall.

BUG=b:29968480
TEST=Ran BubbleUPnP on ryu and minnie connected to the same
wireless AP. Adding this rule results in an entry in the
for the ryu coming up on activity startup on the minnie.
Verified that library browsing and streaming works manually.

Change-Id: I001dcd478c49498db1252fd28cb7fdefab36a393
Reviewed-on: https://chromium-review.googlesource.com/365465
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Ian Wyszynski <wyszynski@google.com>