autotest: Disable kernel_TPMPing until it works more reliably
This test wasn't running on 3.10, 3.14, or 3.18 anyway due to an
incorrect version check.  Then it was failing on 4.4, but because of
missing log spam and not because of a missing feature.  We could fix the
version check and add the log spam to 4.4 but it sounds like folks would
like a better way to check this than extra kernel log spam anyway.  For
now we'll disable.

BUG=chromium:634341
TEST=emerge-${BOARD} -pv autotest-tests-tpm

Change-Id: Icbd1ed3d79e9d84fe490ca3037e0ebf7a4cbd4d2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/366135
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>