arc-mesa-img: Add 2 HACKs for HAL_PIXEL_FORMATs
We don't have proper API to query such buffers for the real internal
format in a generic, scalable way, so let's just hardcode
HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED to RGBX_8888, and
HAL_PIXEL_FORMAT_YCbCr_420_888 to YV12 for the time being,
in the same way as we hardcoded it in drm_gralloc.

TODO: Revert when we have the necessary API in place (see b/29886289).

TEST=android.media.cts.MediaCodecCapabilitiesTest#testAllNonTunneledVideoCodecsSupportFlexibleYUV
TEST=android.view.cts.WindowTest#testSetLocalFocus
BUG=b:28671744
BUG=b:29061386

Change-Id: I7dcb2764820d30795c8eae16221051fe3cf32456
Reviewed-on: https://chromium-review.googlesource.com/367460
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>